client-windows/requirements.txt
openai>=1.0.0
mss>=9.0.1
pillow>=10.0.0
colorama>=0.4.6
numpy